+++ spec/classes/golang_spec.rb
@@ -3,6 +3,11 @@
 require 'spec_helper'
 
 describe 'golang' do
+  let(:pre_condition) { <<~'PUPPET' }
+    # Overide golang::latest_version to avoid dependence on https://go.dev.
+    function golang::latest_version($_url) { '1.0.0' }
+  PUPPET
+
   on_supported_os.each do |os, os_facts|
     context "on #{os}" do
       let(:facts) { os_facts }